fix: Remove Fragment dependency in AsyncTask
As it is actually a bad practice passing a Fragment instance (or even
Activity instances) into AsyncTask. Chances are, if for some reason
the fragment is destroyed, we'll get a NullPointerException Instead,
pass a listener implemented by the caller of the AsyncTask and add
a null check to it before calling it's functions.